Реализовался MIgration, Реализовался все методы для работы с реальной БД

+++ RazorPagesGeneral/Pages/Employess/Edit.cshtml.cs
@@ -1,7 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using RazorPagesProject.Servicess;
@@ -52,7 +49,10 @@ namespace RazorPagesGeneral.Pages.Employess
                     if (Employee.PhotoPath != null)
                     {
                         string filePath = Path.Combine(_webHostEnvironment.WebRootPath, "images", Employee.PhotoPath);
-                        System.IO.File.Delete(filePath);
+                        
+                        if(Employee.PhotoPath!="noimage.png")
+                            System.IO.File.Delete(filePath);
+
                     }
 
                     Employee.PhotoPath = ProcessUploadFile();
